Publicize the removed iter type with docs, move SystemParam impl for  to removal detection file
